checkboxes changed and functional, desktops styling fixed, deleted files i no longer need and indentation check

+++ code/src/components/Header.js
@@ -1,7 +1,10 @@
 import React from 'react'
 import { useSelector, useDispatch } from 'react-redux'
+
 import { todos } from '../reducers/todos'
+
 import moment from 'moment'
+import swal from 'sweetalert'
 
 import { Container, Background , Text, DateText, SectionOne, SectionTwo, TaskCount, Button } from '../styling/HeaderStyling'
 
@@ -12,26 +15,35 @@ export const Header = () => {
     const completedItems = items.filter((item) => item.completedTask)
 
     const handleRemoveAll = () => {
-        dispatch(todos.actions.removeAll())
+      swal({
+        title:'Are you sure',
+        text:'Do you really want to delete all todos?',
+        buttons:true, 
+        dangerMode: true,
+      })
+        .then((willDelete) => {
+            if(willDelete) {
+              dispatch(todos.actions.removeAll())
+            }
+        })   
     }
-
     return (
-    <Container>      
-        <Background background ='#14274e'>
-            <SectionOne>
-                <Text>Todo</Text> 
-                <DateText>{moment().format("MMM Do ")} </DateText>
-            </SectionOne>    
-            <SectionTwo>
-                <TaskCount>{completedItems.length} tasks</TaskCount>
-                    <Button
-                        className="remove-all"
-                        type="button"
-                        onClick ={handleRemoveAll}>
-                        Delete All!
-                    </Button>
-            </SectionTwo>
-        </Background >
+    <Container>  
+      <Background background ='#14274e'>
+        <SectionOne>
+          <Text>Todo</Text> 
+            <DateText>{moment().format("MMM Do ")} </DateText>
+        </SectionOne>    
+          <SectionTwo>
+            <TaskCount>{completedItems.length} tasks</TaskCount>
+              <Button
+                className="remove-all"
+                type="button"
+                onClick ={handleRemoveAll}>
+                Delete All!
+              </Button>
+          </SectionTwo>
+      </Background >
     </Container>      
-    )
+  )
 }
